sample_projects/rag_chatbot_streamlit/requirements.txt
langchain==0.2.14
langchain_chroma==0.1.3
langchain_community==0.2.12
langchain_core==0.2.35
langchain_openai==0.1.22
langchain_text_splitters==0.2.2
python-dotenv==1.0.1
streamlit==1.37.1
